Create README.md
In this project I used Control Flow, Booleans, Variables and Strings to create an engaging project. 

In order to make this application as engaging as possible, I declared a variable 'name' and assigned it to the name of the person who will be asking the Magic 8-Ball.

Just in case the asker does not provide a name, and the value of name is an empty string, I created an if/else statement:
> If the name is an empty string, it will only print the question.
> Else, the player’s name and question are both printed.

I also had to ensure that if the 'question' string was empty, and the user does not provide a question, that is would still work. 
To ensure the fabric of reality was safe, I created an if/else statement where: 
> If the question is an empty string, print out a message to the user ("The Magic 8-Ball cannot provide a fortune unless you ask it something.")
> Else, print the name and question, with the Magic 8-Ball’s answer.